+++ crates/core/src/websocket/types.rs
@@ -199,7 +199,7 @@ pub struct InstrumentSubscription {
 impl InstrumentSubscription {
     /// Creates a new subscription entry from typed values.
     // O(1) EXEMPT: begin — subscription constructor, runs at subscribe time, not per tick
-    pub fn new(segment: ExchangeSegment, security_id: u32) -> Self {
+    pub fn new(segment: ExchangeSegment, security_id: u64) -> Self {
         Self {
             exchange_segment: segment.as_str().to_string(),
             security_id: security_id.to_string(),
